requirements.txt
# Core
llama-cpp-python
llama-index
llama-index-llms-llama-cpp
llama-index-embeddings-huggingface
llama-index-vector-stores-qdrant
llama-index-tools-google

# PDF & Processing
pymupdf
pillow
qdrant-client

# Utils
python-dotenv
pydantic
requests
huggingface_hub

# Hardware Detection
psutil>=5.9.0